Individual Floors
Beginning to physically map individual floors, and coming up with a plan for how to navigate within those floors. Starting with LA2 as I believe it will be one of the most common and most complicated floors to map. Once this is done, adding the other rooms is as simple as making and encoding their physical maps.